1. 统一REST和二进制端口，通过判断前4个字节来判断是什么请求，POST（REST只支持POST请求）为REST、KYLN为二进制 2. 修改二进制协议头，增加MAGIC "KYLN"

+++ netty/src/main/java/org/kylin/transport/netty/client/MessageHandler.java
@@ -22,7 +22,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 public class MessageHandler extends ChannelDuplexHandler {
     Logger logger = RpcLogger.getLogger();
 
-    Map<Long, TransportFuture> pending = new HashMap<Long, TransportFuture>();
+    Map<Integer, TransportFuture> pending = new HashMap<Integer, TransportFuture>();
     Client.Listener[] listeners;
 
     public MessageHandler(Client.Listener... listeners) {
