+++ protocol/src/main/java/org/kylin/protocol/message/Mid.java
@@ -1,17 +1,17 @@
 package org.kylin.protocol.message;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Created by jimmey on 15-6-26.
  */
 public class Mid {
 
-    private static final AtomicLong seq = new AtomicLong();
+    private static final AtomicInteger seq = new AtomicInteger();
 
-    public static long next() {
-        long n = seq.incrementAndGet();
-        if (n == Long.MAX_VALUE) {
+    public static int next() {
+        int n = seq.incrementAndGet();
+        if (n == Integer.MAX_VALUE) {
             seq.set(0);
         }
         return n;
